+++ src/lib/base/java/java/lang/System.java
@@ -33,5 +33,5 @@ public class System
 	
     public static native int currentTimeMillis();
     
-    public static native void arraycopy(Object src, int src_position, Object dst, int dst_position, int length);
+    public static native void arraycopy(Object src, short src_position, Object dst, short dst_position, short length);
 }
